@@ -3,16 +3,20 @@
 	"version": "0.0.3",
 	"description": "A collection of icons in the style of Element UI.",
 	"repository": {
-    "type": "git",
-    "url": "https://github.com/richhost/element-extended-icon-pack"
-  },
-  "homepage": "https://icons.xiangsu.fun",
-	"keywords": ["icon", "svg", "svelte", "element-ui"],
+		"type": "git",
+		"url": "https://github.com/richhost/element-extended-icon-pack"
+	},
+	"homepage": "https://icons.xiangsu.fun",
+	"keywords": [
+		"icon",
+		"svg",
+		"svelte",
+		"element-ui"
+	],
 	"author": "Abiee",
 	"license": "MIT",
 	"scripts": {
 		"build": "pnpm copy:license && pnpm build:icons && pnpm package",
-		"build:aaa": "vite build && npm run package",
 		"copy:license": "copy-license --packageName=svelte",
 		"build:icons": "build-icons --templateSrc=./scripts/template.mjs --iconFileExtension=.svelte --output=./src/lib/icons --entryWithFileExtension",
 		"package": "svelte-kit sync && svelte-package && publint",
@@ -32,21 +36,21 @@
 		"!dist/**/*.spec.*"
 	],
 	"peerDependencies": {
-		"svelte": "^4.0.0"
+		"svelte": "^5.0.0-next.1"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.1",
-		"@sveltejs/kit": "^2.0.6",
-		"@sveltejs/package": "^2.2.4",
-		"@sveltejs/vite-plugin-svelte": "^3.0.1",
+		"@sveltejs/adapter-auto": "^3.2.2",
+		"@sveltejs/kit": "^2.5.18",
+		"@sveltejs/package": "^2.3.2",
+		"@sveltejs/vite-plugin-svelte": "^3.1.1",
 		"build-icons": "workspace:*",
 		"copy-license": "workspace:*",
 		"publint": "^0.1.16",
-		"svelte": "^4.2.8",
-		"svelte-check": "^3.6.2",
-		"tslib": "^2.6.2",
-		"typescript": "^5.3.3",
-		"vite": "^5.0.10"
+		"svelte": "5.0.0-next.200",
+		"svelte-check": "^3.8.4",
+		"tslib": "^2.6.3",
+		"typescript": "^5.5.4",
+		"vite": "^5.3.5"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
