@@ -3,7 +3,7 @@
 	import * as icons from '@element-extended-icon-pack/svelte';
 </script>
 
-<FillEleArrowCaretDouble size={20} color="red" />
+<FillEleArrowCaretDouble size="48" color="#FF3E00" />
 
 <div>
 	{#each Object.entries(icons) as [_, icon]}
